Add files via upload
Machine failures in factories lead to heavy losses and sometimes even the demise of workers. The failures that we are
referring to here are the ones regarding the loosely attached parts of any machine that are caused due to negligence by the machine
construction workers or less scrutiny by the supervisors. This can be easily prevented by installing machines more cautiously or
installing more sensors in the factories to detect failures. But the prospects are very limited considering some factors like the budget,
work power, etc. into concern. As the construction industry's working environment is very complex and thousands of workers are
being injured or killed in accidents every year, safety needs to be taken into consideration. The objective of this paper is to explore the
different computer vision techniques that can be used to improve occupational safety.